Fixed KingMoveCalc and BishopMoveCalc to pass tests
My King and bishop classes could not pass the boundaries tests and capture tests. I rewrote the king class to have individual methods for each move, so I could keep track of pieces on the square it moves too. I also fixed the ChessBoard class so it would create piece out of bounds.